refactor(ui/canvas): enhance component organization and introduce new hooks
- Updated `AGENTS.md` to clarify the structure of the `src/components/` directory and added details about TLDraw canvas hooks.
- Refactored `canvas-space.tsx` to utilize new hooks for canvas rehydration, thread reset, and event handling, improving code clarity and maintainability.
- Exported `CanvasLogger` type from `useCanvasComponentStore.ts` for better type accessibility across modules.
